requirements.txt
pandas
numpy
requests
python-dotenv
fredapi
polygon-api-client
newsapi-python
textblob
ccxt
colorama
